Move ExceptionMessage in other package
I move the constants messages in a new package in the exceptions package
to identify more quick where are all the messages when an exception is
throw it
