+++ src/main/java/com/agilethought/internship/sso/exception/errorMessages/PathErrorMessage.java
@@ -1,0 +1,6 @@
+package com.agilethought.internship.sso.exception.errorMessages;
+
+public class PathErrorMessage {
+	
+	public static final String PATH_API = "/api/v1/user/";
+}
