--- src/main/java/com/agilethougth/intership/sso/errorhandling/PathErrorMessage.java
@@ -1,6 +1,0 @@
-package com.agilethougth.intership.sso.errorhandling;
-
-public class PathErrorMessage {
-	
-	public static final String PATH_API = "/api/v1/user/";
-}
